[network] Split TransportManager and PeerManager
Closes: #9032
