+++ network/src/peer_manager/types.rs
@@ -1,0 +1,80 @@
+// Copyright (c) The Diem Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+use crate::{
+    peer::DisconnectReason,
+    peer_manager::PeerManagerError,
+    protocols::{
+        direct_send::Message,
+        rpc::{InboundRpcRequest, OutboundRpcRequest},
+    },
+    transport::{Connection, ConnectionMetadata},
+};
+use diem_config::network_id::NetworkContext;
+use diem_types::{network_address::NetworkAddress, PeerId};
+use futures::channel::oneshot;
+use serde::Serialize;
+use std::{fmt, sync::Arc};
+
+/// Request received by PeerManager from upstream actors.
+#[derive(Debug, Serialize)]
+pub enum PeerManagerRequest {
+    /// Send an RPC request to a remote peer.
+    SendRpc(PeerId, #[serde(skip)] OutboundRpcRequest),
+    /// Fire-and-forget style message send to a remote peer.
+    SendDirectSend(PeerId, #[serde(skip)] Message),
+}
+
+/// Notifications sent by PeerManager to upstream actors.
+#[derive(Debug)]
+pub enum PeerManagerNotification {
+    /// A new RPC request has been received from a remote peer.
+    RecvRpc(PeerId, InboundRpcRequest),
+    /// A new message has been received from a remote peer.
+    RecvMessage(PeerId, Message),
+}
+
+#[derive(Debug, Serialize)]
+pub enum ConnectionRequest {
+    DialPeer(
+        PeerId,
+        NetworkAddress,
+        #[serde(skip)] oneshot::Sender<Result<(), PeerManagerError>>,
+    ),
+    DisconnectPeer(
+        PeerId,
+        #[serde(skip)] oneshot::Sender<Result<(), PeerManagerError>>,
+    ),
+}
+
+#[derive(Clone, PartialEq, Serialize)]
+pub enum ConnectionNotification {
+    /// Connection with a new peer has been established.
+    NewPeer(ConnectionMetadata, Arc<NetworkContext>),
+    /// Connection to a peer has been terminated. This could have been triggered from either end.
+    LostPeer(ConnectionMetadata, Arc<NetworkContext>, DisconnectReason),
+}
+
+impl fmt::Debug for ConnectionNotification {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self)
+    }
+}
+
+impl fmt::Display for ConnectionNotification {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ConnectionNotification::NewPeer(metadata, context) => {
+                write!(f, "[{},{}]", metadata, context)
+            }
+            ConnectionNotification::LostPeer(metadata, context, reason) => {
+                write!(f, "[{},{},{}]", metadata, context, reason)
+            }
+        }
+    }
+}
+
+#[derive(Debug, Serialize)]
+pub enum TransportNotification<TSocket> {
+    NewConnection(#[serde(skip)] Connection<TSocket>),
+    Disconnected(ConnectionMetadata, DisconnectReason),
+}
